requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.2.0
requests==2.32.3
